Added OpenWeather Api service

+++ app/src/main/java/ru/devsoland/meteowashkotlin/data/OWCResponse/Wind.kt
@@ -1,6 +1,6 @@
 package ru.devsoland.meteowashkotlin.data.OWCResponse
 
 data class Wind(
-        val deg: Float,
-        val speed: Float
+        val deg: Double,
+        val speed: Double
 )
